[ artf498817] fixed jenkins html and java  build failure, deploy location log info in all plugins and selected db sql file showup in deploy pop up

+++ plugins/drupal-maven-plugin/src/main/java/com/photon/phresco/plugins/DrupalDeploy.java
@@ -23,11 +23,8 @@ import java.io.BufferedReader;
 import java.io.File;
 import java.io.IOException;
 import java.io.InputStreamReader;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
-import org.apache.commons.collections.CollectionUtils;
 import org.apache.maven.plugin.AbstractMojo;
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.project.MavenProject;
@@ -203,8 +200,10 @@ public class DrupalDeploy extends AbstractMojo implements PluginConstants {
 				break;
 			}
 			File deployDir = new File(deployLocation);
-			FileUtils.mkdir(deployDir.getPath().trim());
-			getLog().info("Project is deploying.........");
+			if (!deployDir.exists()) {
+				throw new MojoExecutionException(" Deploy Directory" + deployLocation + " Does Not Exists ");
+			}
+			getLog().info("Project is deploying into " + deployLocation);
 			FileUtils.copyDirectoryStructure(tempDir.getParentFile(), deployDir);
 			getLog().info("Project is deployed successfully");
 		} catch (Exception e) {
